fix: Return lsp config data in authenticated response
Before we have been bombarding our api everytime flutter decides to re-render a screen. I presume this is the reason, why we run into a timeout every once in a while, and why it took some time to load.

The lsp config data is now fetched once at authentication and then cached on the app side. This will significantly reduce the amount of external api calls.